refactor: improve test structure and cleanup
