+++ backend/pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = core.settings
 python_files = test_*.py
+addopts = --ds=core.settings
